# backend/docs/openapi.yaml
openapi: '3.0.3'
info:
  title: Sistem Akademik Kampus Terintegrasi - Backend API
  description: |
    **API Backend MVP 1** untuk Sistem Akademik Kampus Terintegrasi (Portal, SIA, SPADA/LMS, PMB).

    Dibangun berdasarkan spesifikasi [SRS.md](../SRS.md) — mencakup Bab 3, 5, 9, 10, 11, 14, 15, 19, 28, 31, 32, 34, 35, dan 39.

    ## Cakupan MVP 1
    - Autentikasi multi-kredensial & manajemen sesi JWT
    - Otorisasi berbasis 8 role aktor (RBAC)
    - Manajemen profil mandiri
    - Dasbor portal kontekstual per role
    - Master data kelembagaan, akademik, fasilitas & kalender
    - Manajemen pengguna & hak akses
    - Jejak audit (Audit Trail)

    ## Autentikasi
    Menggunakan **Bearer Token (JWT)**. Kirim header `Authorization: Bearer <accessToken>` pada setiap endpoint yang memerlukan autentikasi.

    ## Standar Respons
    Semua respons mengikuti format:
    ```json
    {
      "success": true|false,
      "message": "Deskripsi hasil operasi",
      "data": { ... },
      "meta": { "page": 1, "limit": 10, "total": 100, "totalPages": 10 }
    }
    ```

    ## Endpoint Alias Konseptual (SRS Bab 32)
    | Alias | Redirect Ke |
    |---|---|
    | `POST /api/login` | `POST /api/v1/auth/login` |
    | `POST /api/logout` | `POST /api/v1/auth/logout` |
    | `GET /api/profile` | `GET /api/v1/profile` |
    | `GET /api/calendar` | `GET /api/v1/calendar` |
    | `GET /api/dashboard` | `GET /api/v1/portal/dashboard` |
...
  version: 1.0.0
  contact:
...
  license:
    name: MIT
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access Token JWT (masa berlaku 15 menit)
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          type: object
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        errors:
          type: array
          items:
            type: object
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        total:
          type: integer
          example: 50
        totalPages:
          type: integer
          example: 5

    # ── Auth ──
...
  responses:
    Unauthorized:
      description: Token JWT tidak valid atau tidak disertakan
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Akses ditolak. Token tidak valid atau tidak ditemukan.
    Forbidden:
      description: Tidak memiliki hak akses untuk operasi ini
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Akses ditolak. Anda tidak memiliki izin yang diperlukan.
...
paths:
...
  /api/v1/krs/saya:
    get:
      tags: [KRS]
      summary: KRS milik mahasiswa pada semester tertentu
      operationId: getMyKrs
      security:
        - bearerAuth: []
      parameters:
        - name: semester_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Draft/KRS mahasiswa, atau null jika periode tidak aktif
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/krs/tersedia:
    get:
      tags: [KRS]
      summary: Katalog kelas yang berhak diambil
      operationId: getKelasTersedia
      security:
        - bearerAuth: []
      parameters:
        - name: semester_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Daftar kelas eligible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v1/krs/saya/submit:
    post:
      tags: [KRS]
      summary: Ajukan KRS untuk persetujuan Dosen Wali
      operationId: submitKrs
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                semester_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: KRS berhasil diajukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          description: Pengajuan ditolak dengan alasan terstruktur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...